# go-ordering/docs/swagger.yaml
definitions:
  model.Menu:
    properties:
      category:
        description: 메뉴 카테고리
        type: string
      countryOf:
        description: 원산지
        type: string
      isDisabled:
        description: 판매여부
        type: boolean
      maxCount:
        description: 판매 가능 갯수
        type: integer
      menuID:
        description: 메뉴 ID
        type: string
      menuName:
        description: 메뉴 이름
        type: string
      popularity:
        description: 인기도
        type: integer
      price:
        description: 가격
        type: integer
      sellerID:
        description: 판매자 ID
        type: string
      sellerName:
        description: 판매자 이름
        type: string
      spicy:
        description: 맵기
        type: string
      status:
        description: 주문 가능 상태 (주문가능, 오늘판매완료, 1개남음)
        type: string
      todayMenu:
        description: 오늘의 추천메뉴 여부
        type: boolean
    type: object
  model.OrdererMenuLink:
    properties:
      menuID:
        description: 메뉴 ID
        type: string
      menuName:
        description: 메뉴이름
        type: string
      orderComment:
        description: 후기
        type: string
      orderNo:
        description: 주문번호
        type: string
      orderStarGrade:
        description: 평점
        type: integer
      orderStatus:
        description: 주문상태 (주문확인중 - 조리중 - 배달중 - 배달완료 - 주문취소)
        type: string
      ordererAddress:
        description: 주문자 주소
        type: string
      ordererID:
        description: 주문자ID
        type: string
      ordererPhone:
        description: 주문자 폰번호
        type: integer
      sellerID:
        description: 판매자 ID
        type: string
    type: object
info:
  contact: {}
  description: 띵동주문이요, 온라인 주문 시스템(Online Ordering System)
  title: WBA [Backend Final Project]
  version: "1.0"
paths:
  /oos/order/changeOrder:
    post:
      consumes:
      - application/json
      description: ChangeOrder 주문 변경 - 주문자 (수정/취소)
      parameters:
      - description: 주문변경 커멘드
        enum:
        - 주문추가
        - 주문취소
        - 정보변경
        in: path
        name: ChangeOrderCmd
        required: true
        type: string
      - description: 주문번호
        in: path
        name: OrderNo
        required: true
        type: string
      - description: 주문자 주소
        in: path
        name: OrdererAddress
        type: string
      - description: 주문자 폰번호
        in: path
        name: OrdererPhone
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrdererMenuLink'
      summary: call ChangeOrder, return ok by json.
  /oos/order/createReview:
    post:
      consumes:
      - application/json
      description: CreateReview 리뷰 등록 - 주문자
      parameters:
      - description: 주문번호
        in: path
        name: OrderNo
        required: true
        type: string
      - description: 후기
        in: path
        name: OrderComment
        required: true
        type: string
      - description: 평점
        in: path
        name: OrderStarGrade
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrdererMenuLink'
      summary: call CreateReview, return ok by json.
  /oos/order/newOrder:
    post:
      consumes:
      - application/json
      description: NewOrder 주문 등록 - 주문자
      parameters:
      - description: 메뉴 ID
        in: path
        name: MenuID
        required: true
        type: string
      - description: 메뉴이름
        in: path
        name: MenuName
        required: true
        type: string
      - description: 주문자 ID
        in: path
        name: OrdererID
        required: true
        type: string
      - description: 주문자 주소
        in: path
        name: OrdererAddress
        type: string
      - description: 주문자 폰번호
        in: path
        name: OrdererPhone
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrdererMenuLink'
      summary: call NewOrder, return ok by json.
  /oos/order/searchMenu:
    post:
      consumes:
      - application/json
      description: SearchMenu 메뉴 검색 - 주문자, 피주문자
      parameters:
      - description: 메뉴이름
        in: path
        name: MenuName
        required: true
        type: string
      - description: 가격
        in: path
        name: Price
        required: true
        type: integer
      - description: 원산지
        enum:
        - 한국
        - 일본
        - 중국
        in: path
        name: CountryOf
        required: true
        type: string
      - description: 메뉴 카테고리
        enum:
        - 한식
        - 일식
        - 중식
        in: path
        name: Category
        required: true
        type: string
      - description: 주문 가능 상태
        enum:
        - 준비중
        - 판매중
        in: path
        name: Status
        required: true
        type: string
      - description: 맵기
        enum:
        - 아주매움
        - 매움
        - 보통
        - 순한맛
        in: path
        name: Spicy
        required: true
        type: string
      - description: 오늘의 추천메뉴 여부
        in: path
        name: TodayMenu
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Menu'
            type: array
      summary: call SearchMenu, return ok by json.
  /oos/order/searchOrder:
    post:
      consumes:
      - application/json
      description: SearchOrder 주문 내역 조회 기능 - 주문자
      parameters:
      - description: 주문자 ID
        in: path
        name: OrdererID
        required: true
        type: string
      - description: 메뉴이름
        in: path
        name: MenuName
        required: true
        type: string
      - description: 주문 상태
        enum:
        - 주문확인중 - 조리중 - 배달중 - 배달완료 - 주문취소
        in: path
        name: OrderStatus
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.OrdererMenuLink'
            type: array
      summary: call SearchOrder, return ok by json.
  /oos/order/searchTodayMenu:
    post:
      consumes:
      - application/json
      description: SearchTodayMenu 오늘의 추천메뉴 리스트
      parameters:
      - description: 판매자 ID
        in: path
        name: SellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Menu'
            type: array
      summary: call SearchTodayMenu, return ok by json.
  /oos/order/viewMenu:
    post:
      consumes:
      - application/json
      description: ViewMenu 메뉴 상세 - 주문자, 피주문자
      parameters:
      - description: 메뉴 ID
        in: path
        name: MenuID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Menu'
      summary: call ViewMenu, return ok by json.
  /oos/order/viewOrder:
    post:
      consumes:
      - application/json
      description: ViewOrder 주문 상세 - 주문자, 피주문자
      parameters:
      - description: 주문번호
        in: path
        name: OrderNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrdererMenuLink'
      summary: call ViewOrder, return ok by json.
  /oos/seller/OrderStatus:
    post:
      consumes:
      - application/json
      description: OrderStatus 주문 내역 조회 - 피주문자
      parameters:
      - description: 판매자 ID
        in: path
        name: SellerID
        required: true
        type: string
      - description: 주문 상태
        enum:
        - 주문확인중 - 조리중 - 배달중 - 배달완료 - 주문취소
        in: path
        name: OrderStatus
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.OrdererMenuLink'
            type: array
      summary: call OrderStatus, return ok by json.
  /oos/seller/createMenu:
    post:
      consumes:
      - application/json
      description: CreateMenu 메뉴 등록 - 피주문자
      parameters:
      - description: 판매자 ID
        in: path
        name: SellerID
        required: true
        type: string
      - description: 판매자 이름
        in: path
        name: SellerName
        required: true
        type: string
      - description: 메뉴이름
        in: path
        name: MenuName
        required: true
        type: string
      - description: 주문 가능 상태
        enum:
        - 준비중
        - 판매중
        in: path
        name: Status
        required: true
        type: string
      - description: 판매 가능 갯수
        in: path
        name: MaxCount
        required: true
        type: integer
      - description: 원산지
        enum:
        - 한국
        - 일본
        - 중국
        in: path
        name: CountryOf
        required: true
        type: string
      - description: 가격
        in: path
        name: Price
        required: true
        type: integer
      - description: 맵기
        enum:
        - 아주매움
        - 매움
        - 보통
        - 순한맛
        in: path
        name: Spicy
        required: true
        type: string
      - default: true
        description: 판매여부
        in: path
        name: IsDisabled
        required: true
        type: boolean
      - default: false
        description: 오늘의 추천메뉴 여부
        in: path
        name: TodayMenu
        type: boolean
      - description: 메뉴 카테고리
        enum:
        - 한식
        - 일식
        - 중식
        in: path
        name: Category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Menu'
      summary: call CreateMenu, return ok by json.
  /oos/seller/deleteMenu:
    put:
      consumes:
      - application/json
      description: DeleteMenu 메뉴 삭제 - 피주문자
      parameters:
      - description: 메뉴 ID
        in: path
        name: MenuID
        required: true
        type: string
      - description: 판매여부
        in: path
        name: IsDisabled
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Menu'
      summary: call DeleteMenu, return ok by json.
  /oos/seller/setTodayMenu:
    post:
      consumes:
      - application/json
      description: SetTodayMenu 오늘의 추천메뉴 여부 - 설정 변경
      parameters:
      - description: 메뉴 ID
        in: path
        name: MenuID
        required: true
        type: string
      - description: 오늘의 추천메뉴 여부
        in: path
        name: TodayMenu
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Menu'
      summary: call SetTodayMenu, return ok by json.
  /oos/seller/updateMenu:
    post:
      consumes:
      - application/json
      description: UpdateMenu 메뉴 수정 - 피주문자 (판매자ID와 메뉴ID를 기준으로 메뉴 업데이트)
      parameters:
      - description: 메뉴 ID
        in: path
        name: MenuID
        required: true
        type: string
      - description: 판매자 ID
        in: path
        name: SellerID
        required: true
        type: string
      - description: 판매자 이름
        in: path
        name: SellerName
        required: true
        type: string
      - description: 메뉴이름
        in: path
        name: MenuName
        required: true
        type: string
      - description: 주문 가능 상태
        enum:
        - 준비중
        - 판매중
        in: path
        name: Status
        required: true
        type: string
      - description: 판매 가능 갯수
        in: path
        name: MaxCount
        required: true
        type: integer
      - description: 원산지
        enum:
        - 한국
        - 일본
        - 중국
        in: path
        name: CountryOf
        required: true
        type: string
      - description: 가격
        in: path
        name: Price
        required: true
        type: integer
      - description: 맵기
        enum:
        - 아주매움
        - 매움
        - 보통
        - 순한맛
        in: path
        name: Spicy
        required: true
        type: string
      - default: true
        description: 판매여부
        in: path
        name: IsDisabled
        required: true
        type: boolean
      - default: false
        description: 오늘의 추천메뉴 여부
        in: path
        name: TodayMenu
        type: boolean
      - description: 메뉴 카테고리
        enum:
        - 한식
        - 일식
        - 중식
        in: path
        name: Category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Menu'
      summary: call UpdateMenu, return ok by json.
swagger: "2.0"